Move cli to internal
These interfaces should not have been exported to begin with. Move
them to the internal package.
