+++ certification/internal/cli/openshift.go
