+++ certification/internal/cli/operatorsdk.go
