+++ certification/internal/cli/skopeo.go
